refactor : reviewURL 폴더 정리 (URLGeneratorForm에 사용하는 컴포넌트들 폴더 이동)
- URLGeneratorForm에 사용하는 컴포넌트들을 해당 폴더의 components 폴더로 이동
