+++ frontend/src/components/reviewURL/URLGeneratorForm/components/index.tsx
@@ -1,0 +1,3 @@
+export { default as PasswordField } from './PasswordField';
+export { default as ReviewGroupDataField } from './ReviewGroupDataField';
+export { default as URLGeneratorButton } from './URLGeneratorButton';
